Start processing xml sent over the connection
This just logs open and close elements for now.

This also restructures the XMPPClientSession a bit. Having both the
session and the parser own the tcp stream is a no go. The parser needs
to own it. Perhaps we can have a version of the parser using a RC or
ARC?